[splitruns] Allow real n, new test

+++ include/splitruns.h
@@ -25,9 +25,8 @@ constexpr double splitruns_epsabs = 1e-15;
 /**
  * F(Tobs | n*N). Evaluates F(Tobs | N) exactly, then approximates.
  */
-double runs_split_cumulative(const double Tobs, const unsigned N, const unsigned n, double epsrel=splitruns_epsrel, double epsabs=splitruns_epsabs);
-
-double runs_split_pvalue(const double Tobs, const unsigned N, const unsigned n, double epsrel=splitruns_epsrel, double epsabs=splitruns_epsabs);
+double runs_split_cumulative(const double Tobs, const unsigned N, const double n, double epsrel=splitruns_epsrel, double epsabs=splitruns_epsabs);
+double runs_split_pvalue(const double Tobs, const unsigned N, const double n, double epsrel=splitruns_epsrel, double epsabs=splitruns_epsabs);
 
 /**
  * Compute \Delta correction term.
